+++ module/batch-action/src/Application/Form/Type/BatchActionFilterIdsType.php
@@ -1,0 +1,53 @@
+<?php
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Ergonode\BatchAction\Application\Form\Type;
+
+use Ergonode\BatchAction\Application\Form\Model\BatchActionFilterIdsFormModel;
+use Ergonode\Core\Application\Form\Type\BooleanType;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CollectionType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class BatchActionFilterIdsType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add(
+                'list',
+                CollectionType::class,
+                [
+                    'allow_add' => true,
+                    'allow_delete' => true,
+                    'entry_type' => TextType::class,
+                ]
+            )
+            ->add(
+                'included',
+                BooleanType::class,
+            );
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults(
+            [
+                'translation_domain' => 'batch-action',
+                'data_class' => BatchActionFilterIdsFormModel::class,
+            ]
+        );
+    }
+
+    public function getBlockPrefix(): ?string
+    {
+        return null;
+    }
+}