+++ module/batch-action/src/Application/Validator/AllFilterDisabled.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See license.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Ergonode\BatchAction\Application\Validator;
+
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * @Annotation
+ */
+class AllFilterDisabled extends Constraint
+{
+    public string $message = 'String value can only be `all`.';
+}