+++ module/designer/src/Infrastructure/Persistence/Query/DbalTemplateQuery.php
@@ -143,6 +143,34 @@ class DbalTemplateQuery implements TemplateQueryInterface
     }
 
     /**
+     * {@inheritDoc}
+     */
+    public function findTemplateIdsByProductIds(array $productIds): array
+    {
+        $qb = $this->connection->createQueryBuilder();
+        $productIdsValues = [];
+        foreach ($productIds as $productId) {
+            $productIdsValues[] = $productId->getValue();
+        }
+        $result = $qb->select('DISTINCT template_id')
+            ->where($qb->expr()->in('id', ':productsIds'))
+            ->setParameter(':productsIds', $productIdsValues, Connection::PARAM_STR_ARRAY)
+            ->from(self::PRODUCT_TABLE)
+            ->execute()
+            ->fetchAll(\PDO::FETCH_COLUMN);
+
+        if (false === $result) {
+            $result = [];
+        }
+
+        foreach ($result as &$item) {
+            $item = new TemplateId($item);
+        }
+
+        return $result;
+    }
+
+    /**
      * @return array
      */
     public function autocomplete(