+++ module/importer/src/Domain/Notification/StartImportNotification.php
@@ -11,19 +11,19 @@ namespace Ergonode\Importer\Domain\Notification;
 use Ergonode\Notification\Domain\NotificationInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\UserId;
 use Ergonode\SharedKernel\Domain\Aggregate\ImportId;
+use Ergonode\SharedKernel\Domain\AggregateId;
+use Ergonode\SharedKernel\Domain\AbstractId;
 
 class StartImportNotification implements NotificationInterface
 {
+    private const TYPE = 'import-started';
     private const MESSAGE = 'Import "%import%" started';
 
     private string $message;
 
-    private UserId $userId;
+    private UserId $authorId;
 
-    /**
-     * @var array
-     */
-    private array $parameters;
+    private AggregateId $objectId;
 
     private \DateTime $createdAt;
 
@@ -34,27 +34,12 @@ class StartImportNotification implements NotificationInterface
     {
         $this->createdAt = new \DateTime();
         $this->message = self::MESSAGE;
-        $this->parameters = [
-            '%import%' => $importId->getValue(),
-        ];
+        $this->objectId = $importId;
     }
 
     public function getMessage(): string
     {
         return $this->message;
-    }
-
-    public function getUserId(): UserId
-    {
-        return $this->userId;
-    }
-
-    /**
-     * @return array
-     */
-    public function getParameters(): array
-    {
-        return $this->parameters;
     }
 
     public function getCreatedAt(): \DateTime
@@ -64,6 +49,26 @@ class StartImportNotification implements NotificationInterface
 
     public function getAuthorId(): ?UserId
     {
-        return $this->userId;
+        return $this->authorId;
+    }
+
+    public function getType(): string
+    {
+        return self::TYPE;
+    }
+
+    public function getObjectId(): ?AbstractId
+    {
+        return $this->objectId;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getParameters(): array
+    {
+        return [
+            '%import%' => $this->objectId->getValue(),
+        ];
     }
 }