+++ module/notification/src/Application/Controller/Api/NotificationGridReadAction.php
@@ -119,7 +119,7 @@ class NotificationGridReadAction
         $user = $this->userProvider->provide();
         $language = $user->getLanguage();
         $grid = $this->gridBuilder->build($configuration, $language);
-        $dataSet = $this->factory->create($this->query->getDataSet($user->getId(), $language));
+        $dataSet = $this->factory->create($this->query->getGridQuery($user->getId(), $language));
         $data = $this->gridRenderer->render($grid, $configuration, $dataSet);
 
         return new SuccessResponse($data);